api/requirements.txt
flask~=1.1.1
flask-cors~=3.0.8
flask-restful~=0.3.7
psycopg2-binary~=2.8.3 
